agents/agent4_bos/requirements.txt
fastapi
uvicorn
langchain
langchain-community
qdrant-client
requests
pypdf
python-docx
python-multipart